requirements.txt
absl-py==2.3.1
altair==5.5.0
annotated-types==0.7.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
cryptography==46.0.5
filelock==3.19.1
flatbuffers==25.12.19
fsspec==2025.10.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-generativeai==0.8.6
google-pasta==0.2.0
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h5py==3.14.0
httplib2==0.31.2
huggingface_hub==0.36.2
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
libclang==18.1.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
ml_dtypes==0.5.4
namex==0.1.0
narwhals==2.18.0
numpy==2.0.2
opt_einsum==3.4.0
optree==0.19.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
plotly==6.6.0
proto-plus==1.27.1
protobuf==5.29.6
pyarrow==21.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.36.2
regex==2026.1.15
requests==2.32.5
rich==14.3.3
rpds-py==0.27.1
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
smmap==5.0.3
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
tf_keras==2.20.1
threadpoolctl==3.6.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
Werkzeug==3.1.6
wrapt==2.1.2
zipp==3.23.0
